Refactor Parser Interfaces
Refactors the parser interfaces so that they make more sense and are
reusable. Adds cdif as a submodule, and implements the current cdif
version for the factories module. Cleanup result class so that it not
longer uses a superfluous `std::optional`. Add a single parser to wrap
all the ones named *factory which just makes life easier for the
consumers.